introduce AUTOGROW
this adds an AUTOGROW init hook to resize the root filesystem which
resides on a block device to grow to the device size.
AUTOGROW may be ONCE or ALWAYS. All other values are ignored. If
ALWAYS is specified the script may also be linked to the everyboot.d
hooks directory. Use with caution if an existing appliance is resized.
Its recommend to make a backup snapshot before.

Signed-off-by: Peter Lieven <pl@kamp.de>