disable executability of fence firstboot.d scripts
Somewhat useful behind the scenes / during development. This doesn't actually
change how anything works in practice because firstboot scripts wouldn't be
executed again anyhow, but it does make it easier to understand that the fence
is disabled and makes the disabling of the fence more symmetrical with the way
it is enabled.